pyproject.toml
[tool.poetry]
name = "tamarintrack"
version = "0.1.0"
description = ""
authors = ["Timo"]

[tool.poetry.dependencies]
python = "^3.9"
dvc = "^2.10.0"
hydra-core = "^1.1.1"
pdoc3 = "^0.10.0"
pillow = "^10.0.0"
torch = ">=2.0.0, !=2.0.1" #TODO: This is needed because the poetry package for torch 2.0.1 is broken!
torchvision = "^0.15.1"
transformers = "^4.31.0"
ftfy = "^6.1.1"
matplotlib = "^3.7.2"
lightning = "^2.0.6"
timm = "^0.9.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pre-commit = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
ignore = ["E501"]
select = ["B", "C", "E", "F", "W", "B9", "I", "Q"]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    ".dvc",
]

# Same as Black.
line-length = 88

[tool.ruff.mccabe]
max-complexity = 10
